+++ R/DAG.R
@@ -91,7 +91,7 @@ define_model_set <- function(..., .common = NULL) {
 est_DAG <- function(DAG, data, tree, model = 'lambda', method = 'logistic_MPLE',
                     boot = 0, ...) {
   stopifnot(inherits(DAG, 'DAG'))
-  dots <- list(...)
+  dots <- check_dots(list(...))
   # scale the continous variables
   r <- rownames(data)
   data[sapply(data, is.numeric)] <- lapply(data[sapply(data, is.numeric)], scale)